Mostly Renaming Variables
Version 2020.3.30f1
removed create floor script it was redundant and we already took it off the project;
Added new names for variables in CreateBlocks, changed some to serialize fields, changed publics to privates that didnt need to be public.
commented some code that I felt I could need explaining in the future.

public int currentStep = 0; Is currently
doing nothing in the code so i commented it out, im not sure what it was supposed to do.
resized prefabs
adjusted skybox temporarily aswell as prefabs. deleted prefabs that were unneccesary.