Adjustments to the Culling system
I changed how the colliders kick out the blocks
The big green collider sits 10 behind the player and kills everything it sees,
The other one sits 250 ish behind the player and kills all the floors.
this means i needed to move the spawner to a seperate empty that wasnt our floor that spawned in before other floors.
I still need to cull things that spawn under the ground or fall off, and i am thinking of just sizing all of the blocks down to 1 because the algo is too stupid to not position them into the platform properly.